refactor: moved page specific components
