+++ src/pages/FacialRecognition/components/ImageAnnotator.jsx
