+++ src/pages/FacialRecognition/components/profiles/ProfileCard.jsx
@@ -7,11 +7,11 @@ import {
 } from '@material-ui/core';
 import { makeStyles } from '@material-ui/core/styles';
 import React, { useEffect, useState } from 'react';
-import { graphqlQuery } from '../../graphql';
-import { PROFILE as PROFILE_GQL_Q } from '../../graphql/query';
-import { roundOff } from '../../utils';
-import CroppedImage from '../images/CroppedImage';
-import LinearBarsProgress from '../progress/LinearBarsProgress';
+import { graphqlQuery } from '../../../../graphql';
+import { PROFILE as PROFILE_GQL_Q } from '../../../../graphql/query';
+import { roundOff } from '../../../../utils';
+import CroppedImage from '../../../../components/images/CroppedImage';
+import LinearBarsProgress from '../../../../components/progress/LinearBarsProgress';
 
 const useStyles = makeStyles((theme) => ({
   faceCard: {
